Improved validity checking, saving and restoring of Firefox profiles for all system users

+++ setup.sh
@@ -11,6 +11,9 @@ sudo_execute()
 
 import_config_files()
 {
+	echo "Configuration of custom config files use"
+	echo
+	echo
 	# Import config files into '~/'
 	if ask "Do you want to import my custom configuration files ? (y/n) "; then
 # Set script and user's directories
@@ -81,9 +84,9 @@ ask()
 update_system()
 {
 	# Updating the system
-	echo "Updating package lists and upgrading installed packages...
-	
-	"
+	echo "Updating package lists and upgrading installed packages..."
+	echo
+	echo
 	sudo apt update -y
 	echo
 	sudo apt upgrade -y
@@ -95,7 +98,7 @@ update_system()
 install_essential()
 {
 	# Install essential programs
-	echo "Installing essential programs : bat, tree, git, vim, curl, wget, zsh, build-essential"
+	echo "Installing essential programs : bat, tree, git, vim, curl, wget, zsh, build-essential..."
 	echo
 	echo
 	sudo apt install -y bat tree git vim curl wget zsh build-essential
@@ -106,6 +109,8 @@ install_essential()
 
 add_aliases()
 {
+	echo "Setting up some custom aliases"
+	echo
 	echo
 	if ask "Do you want to use some of my custom aliases (choice possible for each alias) ? (y/n) "; then
 	# Configure personnal aliases
@@ -213,6 +218,9 @@ close it and reopen it to apply changes for next times."
 
 personnalize_terminal()
 {
+	echo "Personalize the terminal's appearence"
+	echo
+	echo
 	if ask "Do you want to use my personal terminal profile ? (background transparency) (y/n) "; then
 	# Configure terminal background transparency
 		echo "Configure terminal background transparency..."
@@ -239,6 +247,9 @@ personnalize_terminal()
 
 install_norminette()
 {
+	echo "Norminette configuration"
+	echo
+	echo
 	if ask "Do you want to install the Holy norminette ? (y/n) "; then
 	# Add 'norminette' module
 		echo "Adding norminette module...
@@ -257,16 +268,68 @@ install_norminette()
 	echo
 }
 
+is_firefox_profile()
+{
+	local dir="$1"
+	[[ -f "$dir/prefs.js" && -f "$dir/places.sqlite" ]]
+}
+
+restore_firefox_profiles()
+{
+	echo "Restoring Firefox profiles for each user in the .deb version..."
+	for user_home in /home/*; do
+		user_name=$(basename "$user_home")
+		profile_target_dir="$user_home/.mozilla/firefox"
+
+		for backup_dir in "$user_home/.mozilla/firefox_backup_"* "$user_home/snap/firefox/common/.mozilla/firefox_backup_"* "$user_home/.var/app/org.mozilla.firefox/.mozilla/firefox_backup_"*; do
+			if [ -d "$backup_dir" ]; then
+				valid_profile_found=false
+				for profile_subdir in "$backup_dir"/*/; do
+					if is_firefox_profile "$profile_subdir"; then
+						echo "Restoring profile of $user_name from $backup_dir..."
+
+# Supprimer uniquement les répertoires de profil existants dans firefox/
+						find "$profile_target_dir" -mindepth 1 -maxdepth 1 -type d -exec rm -rf {} +
+
+# Copier les profils de sauvegarde dans firefox/
+						cp -a "$backup_dir/"* "$profile_target_dir/"
+						echo "Restoration of $user_name's profile complete."
+						valid_profile_found=true
+					fi
+				done
+				if ! $valid_profile_found; then
+					echo "No valid Firefox profile backup found for $user_name in $backup_dir."
+					echo "Contents of $backup_dir: $(ls "$backup_dir")"
+				fi
+			else
+				echo "Backup directory $backup_dir does not exist for $user_name."
+			fi
+		done
+
+		if [ -d "$profile_target_dir" ] && [ "$(ls -A "$profile_target_dir")" ]; then
+			echo "Complete restoration of $user_name's profile to the .deb version."
+		else
+			echo "No profile backup restored for $user_name."
+		fi
+	done
+	echo
+	echo "Profile restoration to .deb version complete for all users."
+	echo "All users can now use Firefox in .deb version with their profiles !"
+}
+
 install_firefox()
 {
+	echo "Firefox in .deb installation"
+	echo
+	echo
 	if ask "Do you want to install Firefox in .deb version ? (y/n) "; then
 	# Install firefox in .deb version (or tranfer the snap firefox version profile
 	# to .deb future installation, and install it)
+		echo
 		echo "Starting Firefox .deb installation..."
 		echo
 # Check disk space for Firefox installation
 		echo "Checking available disk space..."
-		echo
 		AVAILABLE_SPACE=$(df ~/ | grep -vE '^Filesystem' | awk '{print $4}')
 		if [ "$AVAILABLE_SPACE" -lt 1048576 ]; then
 			echo "Not enough disk space for Firefox installation. Installation cancelled."
@@ -292,30 +355,8 @@ install_firefox()
 # Perform the actual installation of Firefox
 			echo "Start of Firefox .deb installation protocol..."
 			echo
-# Verification and copy of snap firefox version profile
-			echo "User profile transfer (from snap/flatpack to .deb)..."
-			echo
-			if [ -d ~/.var/app/org.mozilla.firefox/.mozilla/firefox/ ]; then
-				mkdir -p ~/.mozilla/firefox/
-				cp -a ~/snap/firefox/common/.mozilla/firefox/* ~/.mozilla/firefox/
-				echo "Transfer of flatpack firefox version profile, done !"
-			else
-				echo "No such flatpack firefox version profile finded. Searching for snap one..."
-			fi
-			if [ -d ~/snap/firefox/common/.mozilla/firefox/ ]; then
-				mkdir -p ~/.mozilla/firefox/
-				cp -a ~/snap/firefox/common/.mozilla/firefox/* ~/.mozilla/firefox/
-				echo "Transfer of snap firefox version profile, done !"
-				echo
-			else
-				echo "No such snap firefox version profile finded. Let's continue !"
-				echo
-			fi
-
-
-
-			echo "Save Firefox profiles for each user (for : .deb, snap, flatpack)..."
-
+# Verification and copy of snap, flatpak and/or .deb firefox version profile
+			echo "Save Firefox profiles for each user (for : .deb, snap, flatpak)..."
 			for user_home in /home/*; do
 				user_name=$(basename "$user_home")
 # .deb installation profile
@@ -332,28 +373,16 @@ install_firefox()
 					echo "Profile (snap) of $user_name saved in $backup_dir !"
 					cp -a "$profile_dir_snap" "$backup_dir"
 				fi
-	# Flatpak installation profile
+# Flatpak installation profile
 				profile_dir_flatpak="$user_home/.var/app/org.mozilla.firefox/.mozilla/firefox"
 				if [ -d "$profile_dir_flatpak" ]; then
 					backup_dir="$user_home/.var/app/org.mozilla.firefox/.mozilla/firefox_backup_$(date +%Y%m%d)"
-					echo "Profile (flatpack) of $user_name saved in $backup_dir !"
+					echo "Profile (flatpak) of $user_name saved in $backup_dir !"
 					cp -a "$profile_dir_flatpak" "$backup_dir"
 				fi
 			done
-			echo "Sauvegarde des profils terminée."
-
-
-
-
-
-
-
-
-
-
-
-
-
+			echo "Profile saving complete !"
+			echo
 # Update and install necessary dependencies
 			echo "Install necessary dependencies..."
 			echo
@@ -361,37 +390,53 @@ install_firefox()
 			echo
 			sudo apt-get install -y curl wget apt-transport-https dirmngr ca-certificates
 			echo
-			echo
 			echo "Needed dependencies are now installed !"
 			echo
-
-# Remove snap and/or flatpack version of firefox, and clean up any previous .deb installations
-			echo "Removing snap and flatpack versions of firefox, and clean up any previous .deb installations..."
 			echo
-			if ! sudo snap remove --purge firefox; then
-				echo "Error : Firefox Snap removal failed. Please check manually."
+# Remove snap and/or flatpak version of firefox, and clean up any previous .deb installations
+			echo "Removing potential existing Firefox versions"
+			echo
+			echo "Remove old Firefox snap installation..."
+			if snap list firefox >/dev/null 2>&1; then
+				if ! sudo snap remove --purge firefox; then
+					echo "Error : Firefox snap removal failed. Please check the cause manually."
+				else
+					echo "Firefox snap version removed with success !"
+				fi
 			else
-				echo "Firefox Snap version removed with success !"
+				echo "Snap 'firefox' is not installed."
+				echo "Error : Firefox snap removal failed. Please check the cause manually."
 			fi
 			echo
-			echo "Removing flatpack version..."
-			echo
-			
-			if ! sudo flatpak uninstall --delete-data org.mozilla.firefox; then
-				echo "Error : Firefox Flatpak removal failed. Please check manually."
+			echo "Remove old Firefox flatpak installation..."
+			if command -v flatpak >/dev/null 2>&1; then
+				if flatpak list | grep -q org.mozilla.firefox; then
+					if ! sudo flatpak uninstall --delete-data org.mozilla.firefox; then
+						echo "Error : Firefox flatpak removal failed. Please check the cause manually."
+					else
+						echo "Firefox flatpak version removed with success !"
+					fi
+				else
+						echo "flatpak 'firefox' is not installed."
+						echo "Error : Firefox flatpak removal failed. Please check the cause manually."
+				fi
 			else
-				echo "Firefox Flatpack version removed with success !"
+				echo "Flatpak is not installed on this system. Skipping flatpak removal for Firefox."
+			fi
 			echo
-			
-			if ! sudo apt purge firefox; then
-				echo "Error: Firefox .deb removal failed. Please check manually."
+			echo "Remove old firefox .deb installation..."
+			if dpkg -l | grep -q firefox; then
+				if ! sudo apt purge firefox; then
+					echo "Error : Firefox .deb removal failed. Please check the cause manually."
+				else
+					echo "Previous firefox .deb version removed with success !"
+				fi
 			else
-				echo "Previous firefox .deb version removed with success !"
+				echo ".deb 'firefox' is not installed."
+				echo "Error : Firefox .deb removal failed. Please check the cause manually."
 			fi
 			echo
 			echo
-			echo
-
 # Configure the official Mozilla repository
 			echo "Create a directory to store APT repository keys if it doesn't already exist..."
 			sudo install -d -m 0755 /etc/apt/keyrings
@@ -404,12 +449,13 @@ install_firefox()
 			echo
 			echo
 			echo "Digital fingerprint verification..."
-			if ! sudo -u "$SUDO_USER" gpg -n -q --import --import-options import-show /etc/apt/keyrings/packages.mozilla.org.asc | awk '/pub/ { getline; gsub(/^ +| +$/, ""); if ($0 == "35BAA0B33E9EB396F59CA838C0BA5CE6DC6315A3") { print "The digital fingerprint of the key matches (" $0 ").\nLet'\''s continue!\n"; exit 0 } else { print "Key verification failure: the fingerprint (" $0 ") does not match the one due."; exit 1 } }'; then
+			if ! sudo -u "$SUDO_USER" gpg -n -q --import --import-options import-show /etc/apt/keyrings/packages.mozilla.org.asc | awk '/pub/ { getline; gsub(/^ +| +$/, ""); if ($0 == "35BAA0B33E9EB396F59CA838C0BA5CE6DC6315A3") { print "The digital fingerprint of the key matches (" $0 ").\nLet'\''s continue!"; exit 0 } else { print "Key verification failure: the fingerprint (" $0 ") does not match the one due."; exit 1 } }'; then
 				echo "Digital fingerprint not as expected (35BAA0B33E9EB396F59CA838C0BA5CE6DC6315A3). Firefox installation has been cancelled for security reasons."
 				echo
 				echo
 				return 1
 			fi
+			echo
 			echo
 			echo "Add the Mozilla APT repository to your list of sources ( /etc/apt/sources.list.d/mozilla.list )..."
 			if ! grep -q "^deb .\+https://packages.mozilla.org/apt" /etc/apt/sources.list.d/mozilla.list; then
@@ -429,19 +475,58 @@ Pin-Priority: 1000
 			echo "APT has been correctly configured to give priority to packages from the Mozilla repository. Let's continue !"
 			echo
 			echo
-#			echo '
-#			Unattended-Upgrade::Allowed-Origins:: "packages.mozilla.org:${distro_codename}";
-#			' | sudo tee /etc/apt/apt.conf.d/51unattended-upgrades-firefox
-#			echo "Firefox preferences and automatic updates configured !"
-#			echo
-
 # Install Firefox from official Mozilla repository
 			echo "Update your package list and install the Firefox .deb package..."
 			echo
 			sudo apt update
-			sudo apt install -y firefox
-			echo
-			echo "Firefox installation done !"
+			if sudo apt install -y firefox; then
+				echo
+				echo "Firefox installation done !"
+				echo
+				echo
+				restore_firefox_profiles
+			else
+				echo "Error: Failed to install Firefox .deb version. Please check the cause manually."
+			fi
+#			for user_home in /home/*; do
+#				user_name=$(basename "$user_home")
+#				profile_target_dir="$user_home/.mozilla/firefox"
+#
+# Check and restore the profile from the .deb backup
+#				backup_dir_deb="$user_home/.mozilla/firefox_backup_$(date +%Y%m%d)"
+#				if [ -d "$backup_dir_deb" ]; then
+#					echo "Restore profile (.deb) of $user_name from $backup_dir_deb..."
+# Delete existing files to avoid conflicts
+#					rm -rf "$profile_target_dir"
+#					cp -a "$backup_dir_deb" "$profile_target_dir"
+#					echo "Restoration of $user_name profile (.deb) complete."
+#				fi
+# Verify and restore profile from snap backup
+#				backup_dir_snap="$user_home/snap/firefox/common/.mozilla/firefox_backup_$(date +%Y%m%d)"
+#				if [ -d "$backup_dir_snap" ]; then
+#					echo "Restore profile (snap) of $user_name from $backup_dir_snap"
+# Copy snap profile to target directory
+#					cp -a "$backup_dir_snap"/* "$profile_target_dir"
+#					echo "Restoration of $user_name profile (snap) complete."
+#				fi
+# Verify and restore the profile from the flatpak backup
+#				backup_dir_flatpak="$user_home/.var/app/org.mozilla.firefox/.mozilla/firefox_backup_$(date +%Y%m%d)"
+#				if [ -d "$backup_dir_flatpak" ]; then
+#					echo "Restore profile (flatpak) of $user_name from $backup_dir_flatpak"
+# Copy flatpak profile to target directory
+#					cp -a "$backup_dir_flatpak"/* "$profile_target_dir"
+#					echo "Restoration of $user_name profile (flatpak) complete."
+#				fi
+# Check whether the target directory has been populated
+#				if [ -d "$profile_target_dir" ] && [ "$(ls -A "$profile_target_dir")" ]; then
+#					echo "Complete restoration of the $user_name profile to the .deb version."
+#				else
+#					echo "No profile backup found for $user_name."
+#				fi
+#			done
+#			echo
+#			echo "Profile restoration to .deb version complete."
+			echo "Installation and configuration of Firefox .deb complete !"
 		else
 			echo "Firefox installation cancelled."
 			echo
@@ -456,7 +541,7 @@ Pin-Priority: 1000
 	echo
 	echo
 
-	# Execute these commands below to return to the snap version of firefox
+	# Execute these commands below to return to the snap version of Firefox
 	#sudo rm -f /etc/apt/sources.list.d/mozilla.list
 	#sudo rm -f /etc/apt/preferences.d/mozilla
 	#sudo rm -f /etc/apt/apt.conf.d/51unattended-upgrades-firefox
@@ -469,6 +554,9 @@ Pin-Priority: 1000
 
 install_discord()
 {
+	echo "Discord .deb installation"
+	echo
+	echo
 	if ask "Do you want to install Discord in .deb version ? (y/n) "; then
 	# Install the .deb version of Discord after user confirmation
 	# (if storage space is sufficient and dependencies are compatible)
@@ -552,9 +640,8 @@ below, before validating installation)..."
 			return 0
 		fi
 	else
-		echo "Discord installation cancelled."
+		echo "Discord .deb installation cancelled."
 	fi
-	echo
 	echo
 	echo
 	echo
